Fixed size_t: which as different sizes on different operating systems. I am no longer sending size_t fields over the network but instead enet_uint32. Also some other fixes and improvements. GCC release now runs against vcc19 debug :)

+++ main.cpp
@@ -25,7 +25,7 @@ int main(int argc, char** argv) {
 
 
     if (!strcmp(argv[1], "-s")) {
-        int maxShips = argc == 3 ? std::stoi(argv[2]) : 1000;
+        size_t maxShips = argc == 3 ? std::stoi(argv[2]) : 1000;
         return server(maxShips);
     }
